Add ReadSerialCommand method to read until `\n`...
Add `ReadSerialCommand` method to attempt to read a new-line-terminated
string from the `Serial` stream into the buffer.  If a such a command
string is successfully read from the `Serial` stream, call the
`ProcessSerialInput` method to parse and handle the command.

By separating the reading of the command into the buffer from the
parsing and handling of the command, sub-classes may only override the
`ProcessSerialInput` method.  This allows the handling of custom
commands directly from the buffer, without having to handle reading from
the `Serial` stream.  It also makes it easy for an overridden
`ProcessSerialInput` method to call the base `ProcessSerialInput` method
to attempt to process any of the default commands for which the
sub-class does not implement special-handling.